requirements.txt
# MIDI functionality
python-rtmidi>=1.5.0

# HID device support (equivalent to node-hid)
hidapi>=0.14.0

# WebSocket server support (equivalent to ws)
websockets>=11.0.0

# Additional utilities
typing-extensions>=4.0.0
